django_vue/requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==42.0.2
Django==3.2.23
django-cors-headers==4.1.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
inflection==0.5.1
install==1.3.5
PyJWT==1.7.1
PyMySQL==1.1.0
pytz==2023.4
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
